Move mimir's ShiftedReadPlanner out of Mimir.
